directorios corregidos
